Implementation of the trie - the building block of the lexer (#1)
Trie will hold all the keywords of the chimera language as keys and
corresponding `TokenKind`s with optional rules to decide match
behaviour. Some keywords need additional requirements to be met so that
they can be counted as a match.
E.g.s 
- `impl` keyword need a separator character after it to be counted as a
match. `impl_a_trait` can be a variable thus the need for any
separation.
- for a parsing stand point, single or multiple whitespace(s) do not
affect anything. Thus, we will be extending the match, if the following
characters of any whitespace character is identical. (Since we don't use
regex, and cannot insert all the possible whitespace lengths)

These behaviours are held in the final node of the trie branch that is
matching.

Note: I also had some fuzzer tests separetly, will be including that
module when the lexer matures.